% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bopr.R
\name{online_leraning}
\alias{online_leraning}
\title{Online learning with stream of samples.}
\usage{
online_leraning(object, newdata = NULL, allow.new = TRUE)
}
\arguments{
\item{object}{\code{BOPR} object}

\item{newdata}{a data.frame of predictors.}

\item{allow.new}{allow new levels of current variable}
}
\value{
S3 \code{BOPR} object; a list of consisting of
 \item{beta_matrix}{beta matrix with mean and variance}
 \item{beta}{scaling parameter}
 \item{prior_prob}{prior of initial parametes}
 \item{epsilon}{parameter to apply dynamics}
}
\description{
Online learning with stream of samples.
}

